+++ coverage/sysmon.py
@@ -3,10 +3,9 @@
 
 """Callback functions and support for sys.monitoring data collection."""
 
-# TODO: https://github.com/python/cpython/issues/111963#issuecomment-2386584080
-
 from __future__ import annotations
 
+import dis
 import functools
 import inspect
 import os
@@ -16,14 +15,17 @@ import threading
 import traceback
 
 from dataclasses import dataclass
-from types import CodeType, FrameType
+from types import CodeType
 from typing import (
     Any,
     Callable,
-    TYPE_CHECKING,
+    Iterable,
+    NewType,
+    Optional,
     cast,
 )
 
+from coverage import env
 from coverage.debug import short_filename, short_stack
 from coverage.misc import isolate_module
 from coverage.types import (
@@ -43,17 +45,28 @@ os = isolate_module(os)
 
 # pylint: disable=unused-argument
 
-LOG = False
+# $set_env.py: COVERAGE_LOG_SYSMON - Log sys.monitoring activity
+LOG = bool(int(os.getenv("COVERAGE_LOG_SYSMON", 0)))
 
 # This module will be imported in all versions of Python, but only used in 3.12+
 # It will be type-checked for 3.12, but not for earlier versions.
 sys_monitoring = getattr(sys, "monitoring", None)
 
-if TYPE_CHECKING:
-    assert sys_monitoring is not None
-    # I want to say this but it's not allowed:
-    #   MonitorReturn = Literal[sys.monitoring.DISABLE] | None
-    MonitorReturn = Any
+DISABLE_TYPE = NewType("DISABLE_TYPE", object)
+MonitorReturn = Optional[DISABLE_TYPE]
+DISABLE = cast(MonitorReturn, getattr(sys_monitoring, "DISABLE", None))
+TOffset = int
+
+ALWAYS_JUMPS: set[int] = set()
+RETURNS: set[int] = set()
+
+if env.PYBEHAVIOR.branch_right_left:
+    ALWAYS_JUMPS.update(
+        dis.opmap[name]
+        for name in ["JUMP_FORWARD", "JUMP_BACKWARD", "JUMP_BACKWARD_NO_INTERRUPT"]
+    )
+
+    RETURNS.update(dis.opmap[name] for name in ["RETURN_VALUE", "RETURN_GENERATOR"])
 
 
 if LOG:  # pragma: debugging
@@ -76,7 +89,10 @@ if LOG:  # pragma: debugging
     assert sys_monitoring is not None
 
     short_stack = functools.partial(
-        short_stack, full=True, short_filenames=True, frame_ids=True,
+        short_stack,
+        full=True,
+        short_filenames=True,
+        frame_ids=True,
     )
     seen_threads: set[int] = set()
 
@@ -99,7 +115,10 @@ if LOG:  # pragma: debugging
             # f"{root}-{pid}-{tslug}.out",
         ]:
             with open(filename, "a") as f:
-                print(f"{pid}:{tslug}: {msg}", file=f, flush=True)
+                try:
+                    print(f"{pid}:{tslug}: {msg}", file=f, flush=True)
+                except UnicodeError:
+                    print(f"{pid}:{tslug}: {ascii(msg)}", file=f, flush=True)
 
     def arg_repr(arg: Any) -> str:
         """Make a customized repr for logged values."""
@@ -130,7 +149,9 @@ if LOG:  # pragma: debugging
                     return ret
                 except Exception as exc:
                     log(f"!!{exc.__class__.__name__}: {exc}")
-                    log("".join(traceback.format_exception(exc))) # pylint: disable=[no-value-for-parameter]
+                    if 1:
+                        # pylint: disable=no-value-for-parameter
+                        log("".join(traceback.format_exception(exc)))
                     try:
                         assert sys_monitoring is not None
                         sys_monitoring.set_events(sys.monitoring.COVERAGE_ID, 0)
@@ -157,17 +178,152 @@ else:
         return _decorator
 
 
+class InstructionWalker:
+    """Utility to step through trails of instructions.
+
+    We have two reasons to need sequences of instructions from a code object:
+    First, in strict sequence to visit all the instructions in the object.
+    This is `walk(follow_jumps=False)`.  Second, we want to follow jumps to
+    understand how execution will flow: `walk(follow_jumps=True)`.
+
+    """
+
+    def __init__(self, code: CodeType) -> None:
+        self.code = code
+        self.insts: dict[TOffset, dis.Instruction] = {}
+
+        inst = None
+        for inst in dis.get_instructions(code):
+            self.insts[inst.offset] = inst
+
+        assert inst is not None
+        self.max_offset = inst.offset
+
+    def walk(
+        self, *, start_at: TOffset = 0, follow_jumps: bool = True
+    ) -> Iterable[dis.Instruction]:
+        """
+        Yield instructions starting from `start_at`.  Follow unconditional
+        jumps if `follow_jumps` is true.
+        """
+        seen = set()
+        offset = start_at
+        while offset < self.max_offset + 1:
+            if offset in seen:
+                break
+            seen.add(offset)
+            if inst := self.insts.get(offset):
+                yield inst
+                if follow_jumps and inst.opcode in ALWAYS_JUMPS:
+                    offset = inst.jump_target
+                    continue
+            offset += 2
+
+
+def populate_branch_trails(code: CodeType, code_info: CodeInfo) -> None:
+    """
+    Populate the `branch_trails` attribute on `code_info`.
+
+    Instructions can have a jump_target, where they might jump to next.  Some
+    instructions with a jump_target are unconditional jumps (ALWAYS_JUMPS), so
+    they aren't interesting to us, since they aren't the start of a branch
+    possibility.
+
+    Instructions that might or might not jump somewhere else are branch
+    possibilities.  For each of those, we track a trail of instructions.  These
+    are lists of instruction offsets, the next instructions that can execute.
+    We follow the trail until we get to a new source line.  That gives us the
+    arc from the original instruction's line to the new source line.
+
+    """
+    log(f"populate_branch_trails: {code}")
+    iwalker = InstructionWalker(code)
+    for inst in iwalker.walk(follow_jumps=False):
+        log(f"considering {inst=}")
+        if not inst.jump_target:
+            # We only care about instructions with jump targets.
+            log("no jump_target")
+            continue
+        if inst.opcode in ALWAYS_JUMPS:
+            # We don't care about unconditional jumps.
+            log("always jumps")
+            continue
+
+        from_line = inst.line_number
+        if from_line is None:
+            continue
+
+        def walk_one_branch(
+            start_at: TOffset, branch_kind: str
+        ) -> tuple[list[TOffset], TArc | None]:
+            # pylint: disable=cell-var-from-loop
+            inst_offsets: list[TOffset] = []
+            to_line = None
+            for inst2 in iwalker.walk(start_at=start_at):
+                inst_offsets.append(inst2.offset)
+                if inst2.line_number and inst2.line_number != from_line:
+                    to_line = inst2.line_number
+                    break
+                elif inst2.jump_target and (inst2.opcode not in ALWAYS_JUMPS):
+                    log(
+                        f"stop: {inst2.jump_target=}, "
+                        + f"{inst2.opcode=} ({dis.opname[inst2.opcode]}), "
+                        + f"{ALWAYS_JUMPS=}"
+                    )
+                    break
+                elif inst2.opcode in RETURNS:
+                    to_line = -code.co_firstlineno
+                    break
+            if to_line is not None:
+                log(
+                    f"possible branch from @{start_at}: "
+                    + f"{inst_offsets}, {(from_line, to_line)} {code}"
+                )
+                return inst_offsets, (from_line, to_line)
+            else:
+                log(f"no possible branch from @{start_at}: {inst_offsets}")
+                return [], None
+
+        # Calculate two trails: one from the next instruction, and one from the
+        # jump_target instruction.
+        trails = [
+            walk_one_branch(start_at=inst.offset + 2, branch_kind="not-taken"),
+            walk_one_branch(start_at=inst.jump_target, branch_kind="taken"),
+        ]
+        code_info.branch_trails[inst.offset] = trails
+
+        # Sometimes we get BRANCH_RIGHT or BRANCH_LEFT events from instructions
+        # other than the original jump possibility instruction.  Register each
+        # trail under all of their offsets so we can pick up in the middle of a
+        # trail if need be.
+        for trail in trails:
+            for offset in trail[0]:
+                if offset not in code_info.branch_trails:
+                    code_info.branch_trails[offset] = []
+                code_info.branch_trails[offset].append(trail)
+
 @dataclass
 class CodeInfo:
     """The information we want about each code object."""
 
     tracing: bool
     file_data: TTraceFileData | None
-    # TODO: what is byte_to_line for?
-    byte_to_line: dict[int, int] | None
+    byte_to_line: dict[TOffset, TLineNo] | None
+
+    # Keys are start instruction offsets for branches.
+    # Values are lists:
+    #   [
+    #       ([offset, offset, ...], (from_line, to_line)),
+    #       ([offset, offset, ...], (from_line, to_line)),
+    #   ]
+    #   Two possible trails from the branch point, left and right.
+    branch_trails: dict[
+        TOffset,
+        list[tuple[list[TOffset], TArc | None]],
+    ]
 
 
-def bytes_to_lines(code: CodeType) -> dict[int, int]:
+def bytes_to_lines(code: CodeType) -> dict[TOffset, TLineNo]:
     """Make a dict mapping byte code offsets to line numbers."""
     b2l = {}
     for bstart, bend, lineno in code.co_lines():
@@ -204,9 +360,6 @@ class SysMonitor(Tracer):
         # A list of code_objects, just to keep them alive so that id's are
         # useful as identity.
         self.code_objects: list[CodeType] = []
-        self.last_lines: dict[FrameType, int] = {}
-        # Map id(code_object) -> code_object
-        self.local_event_codes: dict[int, CodeType] = {}
         self.sysmon_on = False
         self.lock = threading.Lock()
 
@@ -230,20 +383,23 @@ class SysMonitor(Tracer):
         assert sys_monitoring is not None
         sys_monitoring.use_tool_id(self.myid, "coverage.py")
         register = functools.partial(sys_monitoring.register_callback, self.myid)
-        events = sys_monitoring.events
+        events = sys.monitoring.events
+
+        sys_monitoring.set_events(self.myid, events.PY_START)
+        register(events.PY_START, self.sysmon_py_start)
         if self.trace_arcs:
-            sys_monitoring.set_events(
-                self.myid,
-                events.PY_START | events.PY_UNWIND,
-            )
-            register(events.PY_START, self.sysmon_py_start)
-            register(events.PY_RESUME, self.sysmon_py_resume_arcs)
-            register(events.PY_RETURN, self.sysmon_py_return_arcs)
-            register(events.PY_UNWIND, self.sysmon_py_unwind_arcs)
+            register(events.PY_RETURN, self.sysmon_py_return)
             register(events.LINE, self.sysmon_line_arcs)
+            if env.PYBEHAVIOR.branch_right_left:
+                register(
+                    events.BRANCH_RIGHT,  # type:ignore[attr-defined]
+                    self.sysmon_branch_either,
+                )
+                register(
+                    events.BRANCH_LEFT,  # type:ignore[attr-defined]
+                    self.sysmon_branch_either,
+                )
         else:
-            sys_monitoring.set_events(self.myid, events.PY_START)
-            register(events.PY_START, self.sysmon_py_start)
             register(events.LINE, self.sysmon_line_lines)
         sys_monitoring.restart_events()
         self.sysmon_on = True
@@ -257,11 +413,7 @@ class SysMonitor(Tracer):
             return
         assert sys_monitoring is not None
         sys_monitoring.set_events(self.myid, 0)
-        with self.lock:
-            self.sysmon_on = False
-            for code in self.local_event_codes.values():
-                sys_monitoring.set_local_events(self.myid, code, 0)
-            self.local_event_codes = {}
+        self.sysmon_on = False
         sys_monitoring.free_tool_id(self.myid)
 
     @panopticon()
@@ -281,23 +433,10 @@ class SysMonitor(Tracer):
         """Return a dictionary of statistics, or None."""
         return None
 
-    # The number of frames in callers_frame takes @panopticon into account.
-    if LOG:
-
-        def callers_frame(self) -> FrameType:
-            """Get the frame of the Python code we're monitoring."""
-            return (
-                inspect.currentframe().f_back.f_back.f_back  # type: ignore[union-attr,return-value]
-            )
-
-    else:
-
-        def callers_frame(self) -> FrameType:
-            """Get the frame of the Python code we're monitoring."""
-            return inspect.currentframe().f_back.f_back  # type: ignore[union-attr,return-value]
-
     @panopticon("code", "@")
-    def sysmon_py_start(self, code: CodeType, instruction_offset: int) -> MonitorReturn:
+    def sysmon_py_start(  # pylint: disable=useless-return
+        self, code: CodeType, instruction_offset: TOffset
+    ) -> MonitorReturn:
         """Handle sys.monitoring.events.PY_START events."""
         # Entering a new frame.  Decide if we should trace in this file.
         self._activity = True
@@ -337,11 +476,16 @@ class SysMonitor(Tracer):
                 file_data = None
                 b2l = None
 
-            self.code_infos[id(code)] = CodeInfo(
+            code_info = CodeInfo(
                 tracing=tracing_code,
                 file_data=file_data,
                 byte_to_line=b2l,
+                branch_trails={},
             )
+            self.code_infos[id(code)] = code_info
+            if self.trace_arcs:
+                populate_branch_trails(code, code_info)
+                log(f"branch_trails for {code}:\n    {code_info.branch_trails}")
             self.code_objects.append(code)
 
             if tracing_code:
@@ -349,90 +493,82 @@ class SysMonitor(Tracer):
                 with self.lock:
                     if self.sysmon_on:
                         assert sys_monitoring is not None
-                        sys_monitoring.set_local_events(
-                            self.myid,
-                            code,
-                            events.PY_RETURN
-                            #
-                            | events.PY_RESUME
-                            # | events.PY_YIELD
-                            | events.LINE,
-                            # | events.BRANCH
-                            # | events.JUMP
-                        )
-                        self.local_event_codes[id(code)] = code
+                        local_events = events.PY_RETURN | events.PY_RESUME | events.LINE
+                        if self.trace_arcs:
+                            assert env.PYBEHAVIOR.branch_right_left
+                            local_events |= (
+                                events.BRANCH_RIGHT  # type:ignore[attr-defined]
+                                | events.BRANCH_LEFT  # type:ignore[attr-defined]
+                            )
+                        sys_monitoring.set_local_events(self.myid, code, local_events)
 
-        if tracing_code and self.trace_arcs:
-            frame = self.callers_frame()
-            self.last_lines[frame] = -code.co_firstlineno
-            return None
-        else:
-            return sys.monitoring.DISABLE
-
-    @panopticon("code", "@")
-    def sysmon_py_resume_arcs(
-        self, code: CodeType, instruction_offset: int,
-    ) -> MonitorReturn:
-        """Handle sys.monitoring.events.PY_RESUME events for branch coverage."""
-        frame = self.callers_frame()
-        self.last_lines[frame] = frame.f_lineno
+        return None
 
     @panopticon("code", "@", None)
-    def sysmon_py_return_arcs(
-        self, code: CodeType, instruction_offset: int, retval: object,
+    def sysmon_py_return(  # pylint: disable=useless-return
+        self,
+        code: CodeType,
+        instruction_offset: TOffset,
+        retval: object,
     ) -> MonitorReturn:
         """Handle sys.monitoring.events.PY_RETURN events for branch coverage."""
-        frame = self.callers_frame()
         code_info = self.code_infos.get(id(code))
         if code_info is not None and code_info.file_data is not None:
-            last_line = self.last_lines.get(frame)
+            assert code_info.byte_to_line is not None
+            last_line = code_info.byte_to_line[instruction_offset]
             if last_line is not None:
                 arc = (last_line, -code.co_firstlineno)
-                # log(f"adding {arc=}")
                 cast(set[TArc], code_info.file_data).add(arc)
-
-        # Leaving this function, no need for the frame any more.
-        self.last_lines.pop(frame, None)
-
-    @panopticon("code", "@", "exc")
-    def sysmon_py_unwind_arcs(
-        self, code: CodeType, instruction_offset: int, exception: BaseException,
-    ) -> MonitorReturn:
-        """Handle sys.monitoring.events.PY_UNWIND events for branch coverage."""
-        frame = self.callers_frame()
-        # Leaving this function.
-        last_line = self.last_lines.pop(frame, None)
-        if isinstance(exception, GeneratorExit):
-            # We don't want to count generator exits as arcs.
-            return
-        code_info = self.code_infos.get(id(code))
-        if code_info is not None and code_info.file_data is not None:
-            if last_line is not None:
-                arc = (last_line, -code.co_firstlineno)
-                # log(f"adding {arc=}")
-                cast(set[TArc], code_info.file_data).add(arc)
-
+                log(f"adding {arc=}")
+        return None
 
     @panopticon("code", "line")
-    def sysmon_line_lines(self, code: CodeType, line_number: int) -> MonitorReturn:
+    def sysmon_line_lines(self, code: CodeType, line_number: TLineNo) -> MonitorReturn:
         """Handle sys.monitoring.events.LINE events for line coverage."""
-        code_info = self.code_infos[id(code)]
-        if code_info.file_data is not None:
+        code_info = self.code_infos.get(id(code))
+        if code_info is not None and code_info.file_data is not None:
             cast(set[TLineNo], code_info.file_data).add(line_number)
-            # log(f"adding {line_number=}")
-        return sys.monitoring.DISABLE
+            log(f"adding {line_number=}")
+        return DISABLE
 
     @panopticon("code", "line")
-    def sysmon_line_arcs(self, code: CodeType, line_number: int) -> MonitorReturn:
+    def sysmon_line_arcs(self, code: CodeType, line_number: TLineNo) -> MonitorReturn:
         """Handle sys.monitoring.events.LINE events for branch coverage."""
         code_info = self.code_infos[id(code)]
-        ret = None
         if code_info.file_data is not None:
-            frame = self.callers_frame()
-            last_line = self.last_lines.get(frame)
-            if last_line is not None:
-                arc = (last_line, line_number)
+            arc = (line_number, line_number)
+            cast(set[TArc], code_info.file_data).add(arc)
+            log(f"adding {arc=}")
+        return DISABLE
+
+    @panopticon("code", "@", "@")
+    def sysmon_branch_either(
+        self, code: CodeType, instruction_offset: TOffset, destination_offset: TOffset
+    ) -> MonitorReturn:
+        """Handle BRANCH_RIGHT and BRANCH_LEFT events."""
+        code_info = self.code_infos[id(code)]
+        added_arc = False
+        if code_info.file_data is not None:
+            dest_info = code_info.branch_trails.get(instruction_offset)
+            log(f"{dest_info = }")
+            if dest_info is not None:
+                for offsets, arc in dest_info:
+                    if arc is None:
+                        continue
+                    if destination_offset in offsets:
+                        cast(set[TArc], code_info.file_data).add(arc)
+                        log(f"adding {arc=}")
+                        added_arc = True
+                        break
+
+        if not added_arc:
+            # This could be an exception jumping from line to line.
+            assert code_info.byte_to_line is not None
+            l1 = code_info.byte_to_line[instruction_offset]
+            l2 = code_info.byte_to_line[destination_offset]
+            if l1 != l2:
+                arc = (l1, l2)
                 cast(set[TArc], code_info.file_data).add(arc)
-            # log(f"adding {arc=}")
-            self.last_lines[frame] = line_number
-        return ret
+                log(f"adding unforeseen {arc=}")
+
+        return DISABLE